Bugs & デザイン

+++ app/controllers/end_user/payments_controller.rb
@@ -21,7 +21,6 @@ class EndUser::PaymentsController < ApplicationController
   def confirm
     @carts = Cart.where(user_id: current_user.id)
     @order = Order.new(order_params)
-
     if @order.shipping_id == nil
       render :new
     else
@@ -31,42 +30,47 @@ class EndUser::PaymentsController < ApplicationController
 
   def complete
     shipping = Shipping.find(params[:shipping])
-    cart_items = Cart.where(user_id: current_user.id)
+    @carts = Cart.where(user_id: current_user.id)
 
-    order = Order.new(
-      how_to_pay: params[:order],
-      delivery_charge: 500,
-      postal_code: shipping.postal_code,
-      prefecture: shipping.prefecture,
-      city: shipping.city,
-      building: shipping.building,
-      phone: shipping.phone,
-      first_name: shipping.first_name,
-      last_name: shipping.last_name,
-      first_kana_name: shipping.first_kana_name,
-      last_kana_name: shipping.last_kana_name,
-      user_id: current_user.id,
-      total_money: calc_total_price(cart_items)
-      # shipping_status:
-    )
-    order.save
-
-    cart_items.each do |cart_item|
-      order_item = OrderItem.new(
-        order_id: order.id,
-        item_id: cart_item.item.id,
-        item_name: cart_item.item.name,
-        jacket_image_id: cart_item.item.jacket_image_id,
-        price: cart_item.item.price,
-        item_count: cart_item.count,
-        artist_name: cart_item.item.artist.name,
-        genre_name: cart_item.item.genre.name,
-        label_name: cart_item.item.label.name
+    if are_sale(@carts)
+      order = Order.new(
+        how_to_pay: params[:order],
+        delivery_charge: 500,
+        postal_code: shipping.postal_code,
+        prefecture: shipping.prefecture,
+        city: shipping.city,
+        building: shipping.building,
+        phone: shipping.phone,
+        first_name: shipping.first_name,
+        last_name: shipping.last_name,
+        first_kana_name: shipping.first_kana_name,
+        last_kana_name: shipping.last_kana_name,
+        user_id: current_user.id,
+        total_money: calc_total_price(@carts)
+        # shipping_status:
       )
-      order_item.save
-      cart_item.destroy
+      order.save
+
+      @carts.each do |cart_item|
+        order_item = OrderItem.new(
+          order_id: order.id,
+          item_id: cart_item.item.id,
+          item_name: cart_item.item.name,
+          jacket_image_id: cart_item.item.jacket_image_id,
+          price: cart_item.item.price,
+          item_count: cart_item.count,
+          artist_name: cart_item.item.artist.name,
+          genre_name: cart_item.item.genre.name,
+          label_name: cart_item.item.label.name
+        )
+        order_item.save
+        cart_item.destroy
+      end
+      flash[:notice] = '注文が完了しました'
+    else
+      flash[:notice] = '売り切れの商品があり、購入できません'
+      render :new
     end
-    flash[:notice] = '注文が完了しました'
   end
 
   private
@@ -80,11 +84,19 @@ class EndUser::PaymentsController < ApplicationController
 
   def calc_total_price(carts)
     sum_price = 0
-
     carts.each do |cart|
         sum_price += cart.item.price * cart.count
     end
-
     return sum_price
+  end
+
+  def are_sale(carts)
+    carts.each do |cart|
+      item = Item.find(cart.item_id)
+      if item.status == 'sold_out'
+        return false
+      end
+    end
+    return true
   end
 end
